requirements.txt
wheel
django
psycopg2
djangorestframework
pandas
openpyxl
SQLAlchemy
django-cors-headers